+++ .gitignore
@@ -6,3 +6,4 @@
 /build
 /captures
 .idea
+.iml
