+++ app/src/androidTest/java/com/medziku/motoresponder/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.medziku.myapplication;
+package com.medziku.motoresponder;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
