+++ app/src/main/java/com/medziku/motoresponder/CellStateCallback.java
@@ -1,0 +1,8 @@
+package com.medziku.motoresponder;
+
+/**
+ * Created by medziku on 22.09.15.
+ */
+public interface CellStateCallback {
+    void onCellState();
+}
