+++ app/src/main/java/com/medziku/motoresponder/LocationUtility.java
@@ -1,18 +1,10 @@
-package com.medziku.myapplication;
+package com.medziku.motoresponder;
 
 
 import android.content.Context;
 import android.location.*;
 import android.os.Bundle;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.Locale;
-
-interface LocationChangedCallback {
-    void onLocationChange(Location location);
-
-}
 
 //interface LocationCityChangedCallback {
 //
