+++ app/src/main/java/com/medziku/motoresponder/LockStateCallback.java
@@ -1,0 +1,8 @@
+package com.medziku.motoresponder;
+
+/**
+ * Created by medziku on 22.09.15.
+ */
+public interface LockStateCallback {
+    void onChangeState(boolean isLocked);
+}
