+++ app/src/main/java/com/medziku/motoresponder/MainActivity.java
@@ -1,4 +1,4 @@
-package com.medziku.myapplication;
+package com.medziku.motoresponder;
 
 import android.app.Activity;
 import android.content.Context;
