+++ app/src/main/java/com/medziku/motoresponder/SMSReceivedCallback.java
@@ -1,0 +1,8 @@
+package com.medziku.motoresponder;
+
+/**
+ * Created by medziku on 22.09.15.
+ */
+public interface SMSReceivedCallback {
+    void onSMSReceived(String phoneNumber, String message);
+}
