+++ app/src/main/java/com/medziku/motoresponder/SendSMSCallback.java
@@ -1,0 +1,10 @@
+package com.medziku.motoresponder;
+
+/**
+ * Created by medziku on 22.09.15.
+ */
+public interface SendSMSCallback {
+    void onSMSSent(String status);
+
+    void onSMSDelivered(String status);
+}
